fix(async_hooks): satisfy repository integration gates
