+++ crates/perry-runtime/src/object/field_get_set/get_field_by_name_tail.rs
@@ -22,23 +22,8 @@ pub(crate) fn get_field_by_name_object_tail(
                 // undefined/null tag or null pointer — return undefined
                 return JSValue::undefined();
             }
-            if !key.is_null() {
-                unsafe {
-                    let key_ptr =
-                        (key as *const u8).add(std::mem::size_of::<crate::StringHeader>());
-                    let key_len = (*key).byte_len as usize;
-                    if let Ok(name) =
-                        std::str::from_utf8(std::slice::from_raw_parts(key_ptr, key_len))
-                    {
-                        if let Some(value) =
-                            crate::async_hooks::try_async_resource_property_dispatch(
-                                raw as i64, name,
-                            )
-                        {
-                            return JSValue::from_bits(value.to_bits());
-                        }
-                    }
-                }
+            if let Some(value) = async_resource_property(raw, key) {
+                return value;
             }
             // Issue #340: small-handle receivers (raw < 0x100000) come
             // from native modules (axios, fastify, ioredis, ...) that
@@ -130,18 +115,8 @@ pub(crate) fn get_field_by_name_object_tail(
     if obj.is_null() {
         return JSValue::undefined();
     }
-    if !key.is_null() {
-        unsafe {
-            let key_ptr = (key as *const u8).add(std::mem::size_of::<crate::StringHeader>());
-            let key_len = (*key).byte_len as usize;
-            if let Ok(name) = std::str::from_utf8(std::slice::from_raw_parts(key_ptr, key_len)) {
-                if let Some(value) =
-                    crate::async_hooks::try_async_resource_property_dispatch(obj as i64, name)
-                {
-                    return JSValue::from_bits(value.to_bits());
-                }
-            }
-        }
+    if let Some(value) = async_resource_property(obj, key) {
+        return value;
     }
     // Same handle-receiver path for already-stripped pointers — happens
     // when the codegen passes a raw i64 handle through the slow path.
