+++ crates/perry-runtime/src/object/field_get_set/ic_miss.rs
@@ -482,7 +482,7 @@ pub extern "C" fn js_object_get_field_ic_miss(
     // (0x200_0000_0000 in is_valid_obj_ptr) masked this; Linux's is 0x1000.
     if !key.is_null() {
         unsafe {
-            let key_ptr = (key as *const u8).add(std::mem::size_of::<crate::StringHeader>());
+            let key_ptr = crate::string::string_data(key);
             let key_len = (*key).byte_len as usize;
             if let Ok(name) = std::str::from_utf8(std::slice::from_raw_parts(key_ptr, key_len)) {
                 if let Some(value) =
