+++ scripts/thread_local_cold_allowlist.json
@@ -1,6 +1,6 @@
 {
   "_comment": "Files still declaring raw `thread_local!`. Every entry is a declaration that pays `_tlv_get_addr` on Darwin; the count is a ratchet, so adding one to an already-listed file fails too. New code should use `crate::perry_thread_local!` \u2014 see crates/perry-runtime/src/tls_hot.rs. Regenerate with scripts/check_thread_locals.py --update.",
-  "_hot_declarations": 227,
+  "_hot_declarations": 256,
   "files": {
     "crates/perry-runtime/src/agent.rs": 1,
     "crates/perry-runtime/src/arena/block.rs": 2,
@@ -24,7 +24,6 @@
     "crates/perry-runtime/src/eh_walker.rs": 1,
     "crates/perry-runtime/src/error.rs": 2,
     "crates/perry-runtime/src/event_pump.rs": 1,
-    "crates/perry-runtime/src/fs/callbacks.rs": 1,
     "crates/perry-runtime/src/fs/dir_glob_watch/watch.rs": 1,
     "crates/perry-runtime/src/fs/filehandle.rs": 1,
     "crates/perry-runtime/src/fs/mod.rs": 1,
@@ -69,6 +68,7 @@
     "crates/perry-runtime/src/node_submodules/test.rs": 1,
     "crates/perry-runtime/src/node_submodules/test_once_unit_tests.rs": 1,
     "crates/perry-runtime/src/node_submodules/test_property.rs": 1,
+    "crates/perry-runtime/src/node_submodules/test_runner.rs": 1,
     "crates/perry-runtime/src/node_submodules/trace_events.rs": 1,
     "crates/perry-runtime/src/object/native_module/callable_exports.rs": 2,
     "crates/perry-runtime/src/object/spill.rs": 1,
